+++ iModelCore/ECDb/ECDb/IssueReporter.cpp
@@ -738,6 +738,7 @@ IssueId ECDbIssueId::ECDb_0738 = IssueId("ECDb_0738");
 IssueId ECDbIssueId::ECDb_0739 = IssueId("ECDb_0739");
 IssueId ECDbIssueId::ECDb_0740 = IssueId("ECDb_0740");
 IssueId ECDbIssueId::ECDb_0741 = IssueId("ECDb_0741");
+IssueId ECDbIssueId::ECDb_0742 = IssueId("ECDb_0742");
 
 //---------------------------------------------------------------------------------------
 // @bsimethod