+++ iModelCore/ECDb/ECDb/IssueReporter.h
@@ -759,6 +759,7 @@ struct ECDB_EXPORT ECDbIssueId
     static ECN::IssueId ECDb_0739;  // this issue id is being used to report instance query related ambiguity
     static ECN::IssueId ECDb_0740;
     static ECN::IssueId ECDb_0741;
+    static ECN::IssueId ECDb_0742;
     };
 
 //---------------------------------------------------------------------------------------